i2c: sunxi: Add sunxi Reduced Serial Bus (RSB) support
The RSB controller looks like an SMBus controller which only supports byte
and word data transfers. It can also do double-word data transfers, but the
i2c subsystem does not support this, nor have we seen devices using this.
The RSB differs from standard SMBus protocol on several aspects:
- it uses addresses set at runtime to address slaves. Runtime addresses
  are sent to slaves using their 12bit hardware addresses. Up to 15
  runtime addresses are available.
- it adds a parity bit every 8bits of data
- only one read access is required to read a byte (instead of a write
  followed by a read access in standard SMBus protocol)
- there's no Ack bit after each read access

This means this bus cannot be used to interface with standard SMBus
devices (the only known device to support this interface is the AXP223,
AXP806, AXP809 PMICs and the AC100 codec/RTC).
However the RSB protocol is an extension of P2WI, which was close enough
to SMBus to be integrated in the I2C subsystem.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>