drm/sun4i: Add support for tcons without channel0 to tcon quirks
The tcon's on A80 and later SoCs have 1 channel per tcon, typically
channel 0 for tcon0, supporting RGB/LVDS output for LCD panels, and
channel 1 for tcon1, for TV timings outputs such as HDMI and possible
TV encoders.

Tcons without channel 1 are already supported with the A33 tcon.
This patch adds support for tcons without channel 0, by adding
a new field to the quirks structure, and adding checks where ch0
is used. The dotclock is also bound to channel 0, so it is only
registered if ch0 is present.

This patch also re-sorts the fields in the quirks structure.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>