dt-bindings: net: bluetooth: add support for Realtek Bluetooth chips
This adds the documentation for Bluetooth functionality of the Realtek
RTL8723BS and RTL8723DS.
Both are SDIO wifi chips with an additional Bluetooth module which is
connected via UART to the host.

The regulator supply names were chosen to match the names found in the
datasheet for an RTL8723DS module. These names also match the ones used
for Broadcom modules. For reference, an RTL8723BS modules uses VD33
instead of VBAT.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>
Signed-off-by: Jeremy Cline <jeremy@jcline.org>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
[wens@csie.org: changed "-bluetooth" to "-bt" in compatible strings]
[wens@csie.org: dropped realtek,config-data and reset-gpios]
[wens@csie.org: added regulators, device-wakeup-gpios, and host-wakeup-gpios]
Signed-off-by: Chen-Yu Tsai <wens@csie.org>

---

This was picked off of patchwork [1].

[1] https://patchwork.ozlabs.org/patch/921191/