drm/sun4i: sun8i: Display layer format in layer update debug message
Right now we are outputing various aspects of a display layer when such
a property is updated. This includes the layer ID, display dimensions
and coordinates, its backing buffer address, whether interlace mode is
active, and whether hardware scaling is used. What's missing is the
framebuffer / plane format used.

Add the framebuffer format to the debug messages. This may help people
who are working on userspace drivers or applications.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>