ASoC: sunxi: sun8i-codec: Rework clock widgets and routes
The sun8i-codec hardware has individual clock gates and reset controls
for its various sub-modules, as well as a configurable clock hierarchy.
These bits were modeled as widgets and routes in the driver, but the
routes were not entirely accurate, and the names were either too generic
or had typos.

This patch reworks the clock and reset control widgets and routes so
that a) the names are clearer and b) the hierarchy is correct. The
routes are now groups by level in the clock hierarchy instead of by
module.

Each module's reset control is still tied to their own module clock.
This is a convenience measure to slightly cut down on the number of
routes. The audio interfaces each have left and right widgets, so this
reduces the number of routes used by one, for each audio interface.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>