media: sun6i: Set module clock to 300 MHz
On the A64, the CSI module clock running at the default 600 MHz results
in invalid memory access from the CSI controller. This may manifest as
color streaks within the captured image or aborts caught by the
underlying ARM Trusted Firmware.

The BSP sets a default clock rate of 300 MHz. Testing on the A64
confirms this avoids the issue described above.

The BSP also tries to set the clock rate to multiples of 108 MHz,
depending on the capture resolution. This is not implemented yet.
Whether this is needed or not is also uncertain.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>