ASoC: sunxi: sun8i-codec: Rename "Digital ADC Mixer" to "AIF1 Slot 0 Mixer"
The "Digital ADC Mixer" is actually a digital mixer for the AIF1 slot 0
output. It is not related to the ADC. For an ADC, the mixer would be on
the analog side. This is likely due to the user manual using the term
"ADC" to refer to audio interface output, and the fact that AIF1 slot 0
is the only supported audio interface, and it's connected to the CPU
side.

Rename the widget so it correctly reflects the nature of the hardware.
The related DAPM routes are still wrong, either in the opposite
direction or bogus. This will be fixed next.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>