ASoC: sunxi: sun8i-codec: Fix DAC widget
The "DAC" widget actually represents the digital side of the DAC, and is
more than just a supply.

Rename "DAC" to "DAC digital" to make it slightly clearer, change the
DAPM widget type from supply to DAC, and rework the DAPM routes so the
audio flows through this widget from the DAC mixer.

This necessitates changes in the device tree as well. Previously the
routes in the device tree describe audio going through the AIF directly
to the analog side of the DAC, which bypasses the mixer in between.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>