ASoC: sunxi: sun8i-codec: Fix ADC widget and routes
The "ADC" widget actually represents the digital side of the ADC, and is
more than just a supply.

Rename "ADC" to "ADC digital" to make it slightly clearer, change the
DAPM widget type from supply to ADC, and rework the DAPM routes so the
audio flows through this widget to the AIF mixer, then out the AIF.

This necessitates changes in the device tree as well. Previously the
routes in the device tree describe audio going from the analog side of
the ADC directly to AIF, which bypasses the mixer in between.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>