ASoC: sunxi: sun8i-codec: Rework AIF1 Slot 0 Mixer control names
The mixer controls for AIF1 Slot 0 incorrectly use the term "Digital
ADC" when in fact they control what goes out AIF1 Slot 0, which happens
to be connected to the CPU side DAI. This is likely due to the user
manual using the term "ADC" to refer to audio interface output. However,
having multiple fake ADCs in the system is very confusing.

Rename the controls such that they accurately describe the audio source,
the audio flow direction, and the output, in the form "<source> to AIF1
Slot 0".

This may be a good candidate to convert to DPCM in the future.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>