ASoC: sunxi: sun8i-codec: Rename AIF1 slot 0 output widgets
The AIF1 slot 0 output widgets currently use the term "ADC. This is
likely due to the user manual using the term "ADC" to refer to audio
interface outputs. However, having multiple fake ADCs in the system is
very confusing.

Rename the widgets so they use the term "Out" instead of "ADC". "Out" is
from the codec's viewpoint, not from the system's. Also rewrite the
section comments in the list of widgets to clearly denote audio
interface inputs and outputs, rather than the incorrect usage of "DAC"
and "ADC".

Signed-off-by: Chen-Yu Tsai <wens@csie.org>